% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.polly_operations.R
\name{get_lexicon}
\alias{get_lexicon}
\title{Returns the content of the specified pronunciation lexicon stored in an AWS Region}
\usage{
get_lexicon(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] Name of the lexicon.}
}
\description{
Returns the content of the specified pronunciation lexicon stored in an AWS Region. For more information, see \href{http://docs.aws.amazon.com/polly/latest/dg/managing-lexicons.html}{Managing Lexicons}.
}
\examples{
# Returns the content of the specified pronunciation lexicon stored in an
# AWS Region.
get_lexicon(
  Name = ""
)

}
