% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.polly_operations.R
\name{list_lexicons}
\alias{list_lexicons}
\title{Returns a list of pronunciation lexicons stored in an AWS Region}
\usage{
list_lexicons(
  NextToken = "string"
)
}
\arguments{
\item{NextToken}{An opaque pagination token returned from previous \code{ListLexicons} operation. If present, indicates where to continue the list of lexicons.}
}
\description{
Returns a list of pronunciation lexicons stored in an AWS Region. For more information, see \href{http://docs.aws.amazon.com/polly/latest/dg/managing-lexicons.html}{Managing Lexicons}.
}
\examples{
# Returns a list of pronunciation lexicons stored in an AWS Region.
list_lexicons()

}
