% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.polly_package.R
\docType{package}
\name{paws.polly-package}
\alias{paws.polly}
\alias{paws.polly-package}
\title{paws.polly: Amazon Polly}
\description{
Amazon Polly is a web service that makes it easy to
    synthesize speech from text.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
