% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.polly_operations.R
\name{put_lexicon}
\alias{put_lexicon}
\title{Stores a pronunciation lexicon in an AWS Region}
\usage{
put_lexicon(Name, Content)
}
\arguments{
\item{Name}{[required] Name of the lexicon. The name must follow the regular express format [0-9A-Za-z]{1,20}. That is, the name is a case-sensitive alphanumeric string up to 20 characters long.}

\item{Content}{[required] Content of the PLS lexicon as string data.}
}
\description{
Stores a pronunciation lexicon in an AWS Region. If a lexicon with the same name already exists in the region, it is overwritten by the new lexicon. Lexicon operations have eventual consistency, therefore, it might take some time before the lexicon is available to the SynthesizeSpeech operation.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/polly/latest/dg/managing-lexicons.html}{Managing Lexicons}.
}
\section{Accepted Parameters}{
\preformatted{put_lexicon(
  Name = "string",
  Content = "string"
)
}
}

\examples{
# Stores a pronunciation lexicon in an AWS Region.
\donttest{put_lexicon(
  Content = "file://example.pls",
  Name = "W3C"
)}

}
