% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.polly_operations.R
\name{synthesize_speech}
\alias{synthesize_speech}
\title{Synthesizes UTF-8 input, plain text or SSML, to a stream of bytes}
\usage{
synthesize_speech(OutputFormat, Text, VoiceId, LexiconNames = NULL,
  SampleRate = NULL, SpeechMarkTypes = NULL, TextType = NULL,
  LanguageCode = NULL)
}
\arguments{
\item{OutputFormat}{The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg\_vorbis, or pcm. For speech marks, this will be json.

When pcm is used, the content returned is audio/pcm in a signed 16-bit, 1 channel (mono), little-endian format.}

\item{Text}{Input text to synthesize. If you specify \code{ssml} as the \code{TextType}, follow the SSML format for the input text.}

\item{VoiceId}{Voice ID to use for the synthesis. You can get a list of available voice IDs by calling the \href{http://docs.aws.amazon.com/polly/latest/dg/API_DescribeVoices.html}{DescribeVoices} operation.}

\item{LexiconNames}{List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice. For information about storing lexicons, see \href{http://docs.aws.amazon.com/polly/latest/dg/API_PutLexicon.html}{PutLexicon}.}

\item{SampleRate}{The audio frequency specified in Hz.

The valid values for \code{mp3} and \code{ogg_vorbis} are "8000", "16000", and "22050". The default value is "22050".

Valid values for \code{pcm} are "8000" and "16000" The default value is "16000".}

\item{SpeechMarkTypes}{The type of speech marks returned for the input text.}

\item{TextType}{Specifies whether the input text is plain text or SSML. The default value is plain text. For more information, see \href{http://docs.aws.amazon.com/polly/latest/dg/ssml.html}{Using SSML}.}

\item{LanguageCode}{Optional language code for the Synthesize Speech request. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly will use the default language of the bilingual voice. The default language for any voice is the one returned by the \href{https://docs.aws.amazon.com/polly/latest/dg/API_DescribeVoices.html}{DescribeVoices} operation for the \code{LanguageCode} parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.}
}
\description{
Synthesizes UTF-8 input, plain text or SSML, to a stream of bytes. SSML input must be valid, well-formed SSML. Some alphabets might not be available with all the voices (for example, Cyrillic might not be read at all by English voices) unless phoneme mapping is used. For more information, see \href{http://docs.aws.amazon.com/polly/latest/dg/how-text-to-speech-works.html}{How it Works}.
}
\examples{
# Synthesizes plain text or SSML into a file of human-like speech.
synthesize_speech(
  LexiconNames = list(
    "example"
  ),
  OutputFormat = "mp3",
  SampleRate = "8000",
  Text = "All Gaul is divided into three parts",
  TextType = "text",
  VoiceId = "Joanna"
)

}
